requirements.txt
Flask==3.0.3
spacy==3.8.2
kubernetes==30.1.0
requests==2.32.3
PyJWT==2.9.0
flask-limiter==3.8.0
fuzzywuzzy==0.18.0
pyspellchecker==0.8.3
torch==2.7.1
sentence-transformers==3.2.0

# --- Compatible LangChain Suite ---
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.8
langchain-huggingface==0.0.3
langchain-chroma==0.1.1
langsmith==0.1.81
# ----------------------------------
